refactor(contato): atualizando tipagem do contato de telefone

+++ database/migrations/2024_07_30_210140_create_contatos.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('contatos', function (Blueprint $table) {
             $table->id();
             $table->string('nome')->nullable();
-            $table->integer('telefone')->nullable();
+            $table->bigInteger('telefone')->nullable();
             $table->timestamps();
         });
     }
